// src/vs/workbench/contrib/LVGL/module/maxgraph/packages/core/src/types.ts
export interface CellHandle {
	state: CellState;
	cursor: string;
	image: ImageBox | null;
	shape: Shape | null;
	active: boolean;
	setVisible: (v: boolean) => void;
	processEvent: (me: InternalMouseEvent) => void;
	positionChanged: () => void;
	execute: (me: InternalMouseEvent) => void;
	reset: () => void;
	redraw: () => void;
	destroy: () => void;
}